fix(tui): fix scroll stuck — usize offset, cap to max_scroll
scroll_offset was u16 and total_rows could overflow on long transcripts.
offset was never capped to max_scroll so scrolling back down was broken.
Now both are usize; offset is capped before computing effective_scroll.

+++ src/tui/app/mod.rs
@@ -49,7 +49,7 @@ pub struct App {
     pub should_quit: bool,
     pub session_id: Option<String>,
     pub connected: bool,
-    pub scroll_offset: u16,
+    pub scroll_offset: usize,
     pub screen: AppScreen,
     /// Tracks when the TUI session started. Used by `/status` to report uptime.
     pub start_time: Instant,
